fix: fetch commits on overview tab when necessary
* fetch commits on overview / commits (if not very recent)
* move overview commits components to proper files
