+++ client/src/project/Project.present.js
@@ -27,20 +27,18 @@
 import React, { Component, Fragment, useState, useEffect } from "react";
 import { Link, Route, Switch } from "react-router-dom";
 import {
-  Container, Row, Col, Alert, DropdownItem, Table, Nav, NavItem, Button, ButtonGroup, Badge,
-  Card, CardBody, CardHeader, Form, FormGroup, FormText, Label, Input, UncontrolledTooltip, ListGroupItem
+  Alert, Button, ButtonGroup, Card, CardBody, CardHeader, Col, Container, DropdownItem, Form, FormGroup,
+  FormText, Input, Label, Row, Table, Nav, NavItem, UncontrolledTooltip
 } from "reactstrap";
-import qs from "query-string";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faStar as faStarRegular } from "@fortawesome/free-regular-svg-icons";
 import {
-  faCodeBranch, faInfoCircle, faStar as faStarSolid,
-  faExclamationTriangle, faLock, faUserFriends, faGlobe, faSearch
+  faCodeBranch, faExclamationTriangle, faUserFriends, faGlobe, faInfoCircle, faLock, faSearch,
+  faStar as faStarSolid,
 } from "@fortawesome/free-solid-svg-icons";
-import { faGitlab } from "@fortawesome/free-brands-svg-icons";
 
 import {
-  Clipboard, ExternalLink, Loader, RenkuNavLink, TimeCaption, RefreshButton, Pagination,
+  Clipboard, ExternalLink, Loader, RenkuNavLink, TimeCaption,
   ButtonWithMenu, InfoAlert, GoBackButton, RenkuMarkdown,
 } from "../utils/UIComponents";
 import { SpecialPropVal } from "../model/Model";
@@ -51,11 +49,9 @@ import { CollaborationList, collaborationListTypeMap } from "../collaboration/li
 import FilesTreeView from "./filestreeview/FilesTreeView";
 import DatasetsListView from "./datasets/DatasetsListView";
 import { ACCESS_LEVELS } from "../api-client";
-import { withProjectMapped } from "./Project";
 import ProjectVersionStatus from "./status/ProjectVersionStatus.present";
 import { NamespaceProjects } from "../namespace";
-import { CommitsView } from "../utils/Commits";
-import { ProjectOverviewStats } from "./overview";
+import { ProjectOverviewCommits, ProjectOverviewStats } from "./overview";
 
 import "./Project.css";
 
@@ -468,18 +464,12 @@ class ProjectViewOverview extends Component {
             <Route exact path={this.props.overviewDatasetsUrl} render={props =>
               <ProjectViewDatasetsOverview {...this.props} />}
             />
-            <Route exact path={this.props.overviewCommitsUrl}
-              render={(props) => {
-                const categories = ["commits", "metadata"];
-                const ProjectViewCommitsConnected = withProjectMapped(ProjectViewCommits, categories);
-                return (
-                  <ProjectViewCommitsConnected
-                    location={props.location}
-                    history={props.history}
-                    projectCoordinator={projectCoordinator}
-                  />
-                );
-              }}
+            <Route exact path={this.props.overviewCommitsUrl} render={props =>
+              <ProjectOverviewCommits
+                location={this.props.location}
+                history={props.history}
+                projectCoordinator={projectCoordinator}
+              />}
             />
             <Route exact path={this.props.overviewStatusUrl} render={props =>
               <Fragment>
@@ -491,116 +481,6 @@ class ProjectViewOverview extends Component {
         </Col>
       </Row>
     </Col>;
-  }
-}
-
-class ProjectViewCommits extends Component {
-  render() {
-    const { commits, metadata } = this.props;
-    const gitlabCommitsUrl = `${metadata.repositoryUrl}/commits`;
-    const tooMany = commits.error && commits.error.message && commits.error.message.startsWith("Cannot iterate more") ?
-      true :
-      false;
-
-    const commitBadgeNumber = `${commits.list.length}${tooMany ? "+" : ""}`;
-    const badge = commits.fetched && !commits.fetching ?
-      (<Badge color="primary">{commitBadgeNumber}</Badge>) :
-      null;
-    const buttonGit = (
-      <Fragment>
-        <ExternalLink
-          role="link"
-          id="commitLink"
-          title={<FontAwesomeIcon icon={faGitlab} />}
-          url={gitlabCommitsUrl}
-          className="text-primary btn ml-2 p-0"
-        />
-        <UncontrolledTooltip placement="top" target="commitLink">
-          Open in GitLab
-        </UncontrolledTooltip>
-      </Fragment>
-    );
-    const info = commits.error && commits.error.message && commits.error.message.startsWith("Cannot iterate more") ?
-      (<ProjectViewCommitsInfo number={commits.list.length} url={gitlabCommitsUrl} />) :
-      null;
-
-    const body = commits.fetching ?
-      (<Loader />) :
-      (<ProjectViewCommitsBody {...this.props} />);
-    return (
-      <Card className="border-0">
-        <CardHeader>
-          Commits {badge}
-          <RefreshButton action={commits.refresh} updating={commits.fetching} message="Refresh commits" />
-          {buttonGit}
-        </CardHeader>
-        <CardBody className="pl-0 pr-0">
-          {body}
-          {info}
-        </CardBody>
-      </Card>
-    );
-  }
-}
-
-class ProjectViewCommitsInfo extends Component {
-  render() {
-    return (
-      <ListGroupItem className="commit-object">
-        <FontAwesomeIcon icon={faInfoCircle} />&nbsp;
-        Cannot load more than {this.props.number} commits. To see the full project history,&nbsp;
-        <ExternalLink role="link" id="commitLink" title="view in GitLab" url={this.props.url} />
-      </ListGroupItem>
-    );
-  }
-}
-
-class ProjectViewCommitsBody extends Component {
-  constructor(props) {
-    super(props);
-    const locationPage = qs.parse(props.location.search);
-    this.state = {
-      currentPage: locationPage.page ? parseInt(locationPage.page) : 1,
-      perPage: 25,
-    };
-  }
-
-  onPageChange(newPage) {
-    this.setState({ currentPage: newPage });
-    const currentSearch = qs.parse(this.props.location.search);
-    const newSearch = qs.stringify({ ...currentSearch, page: newPage });
-    this.props.history.push({ pathname: this.props.location.pathname, search: newSearch });
-  }
-
-  render() {
-    const { commits, metadata } = this.props;
-    const { currentPage, perPage } = this.state;
-
-    if (commits.fetching || !commits.fetched)
-      return <Loader />;
-
-    const firstCommit = (currentPage - 1) * perPage;
-    const lastCommit = currentPage * perPage;
-
-    return (
-      <Fragment>
-        <CommitsView
-          commits={commits.list.slice(firstCommit, lastCommit)}
-          fetched={commits.fetched}
-          fetching={commits.fetching}
-          error={commits.error}
-          urlRepository={metadata.repositoryUrl}
-          urlDiff={`${metadata.repositoryUrl}/commit/`}
-        />
-        <Pagination
-          className="mt-2"
-          currentPage={currentPage}
-          perPage={perPage}
-          totalItems={commits.list.length}
-          onPageChange={this.onPageChange.bind(this)}
-        />
-      </Fragment>
-    );
   }
 }
 
@@ -1437,4 +1317,4 @@ class ProjectView extends Component {
 export default { ProjectView };
 
 // For testing
-export { filterPaths, ProjectViewCommitsBody };
+export { filterPaths };
